refs #43, #47
HYPRE now builds with MPI and SHARED. A bit of a sledgehammer
move to fix this. The env variable CC now contains all
the CFLAGS and LDFLAGS. Went with this because
we'll also run into this problem once we build HYPRE
with external LAPACK/BLAS libraries.

Emailed the HYPRE team about this bug. Still broken in the latest
version 2.9.0b.

Compiler errors in UbikSolve and Truchas Parser. Need to
resolve before closing this ticket.



git-svn-id: truchas/trunk@42