refs #48
Removed the INSTALL_DIR definitions from the ExternalProject_Add 
commands. CMake would not complete processing because it tries to create 
this directory which is bad if the externals are to be installed
in a location other the default.

I have several questions how to proceed to fully implement this.
In the mean time, here's the work around

(1) cmake -D TruchasExternal_INSTALL_PREFIX:PATH=/some/path ....
sets the root path for the external projects.

(2) Process to install with sudo
    o make zlib-build
    o sudo make zlib-install (HDF5 needs to see this to configure)
    o make hdf5-build netcdf-build swig-build hypre-build
    o sudo make hdf5-install netcdf-install swig-install hypre-install
    o make all
    o make sharen (test suite)
    o sudo make install, if CMAKE_INSTALL_PREFIX also requires sudo

HDf5 and NetCDF take considerable time to build, I suggest using '-j'
when building these projects to speed up the build.

At this time, the build does not search for installations. It only
builds the externals. Feature will be implemented soon #49




git-svn-id: truchas/trunk@46