refs #49
Introduce verification process for an external package
For HYPRE, SWIG, HDF5, and NETCDF, each package is searched
if found the package is then verified against requirements.
This verification is defined in the Verify_* files found under the
external/cmake directory. 

ZIB is now only added to the build if HDF5 or NETCDF are built.

Add option ENABLE_ExternalSearch (DEFAULT ON) wich enables the
search. If set to OFF the build will ALWAYS build all the packages.

Limited testing! May expose dependency issues. Fixed a few
of those when packages where removed from the build and that 
exposed dependency problems. 


git-svn-id: truchas/trunk@53