refs #77, #106
Added the verify and find module files. Also changed the 
petaca CMake variables to upper case PETACA. This convention
is handled consistently in CMake. One of the few places
that CMake *is* case sensitive. 

Found errors in the CMake files in directories where 
Petaca is used directly and corrected with the new find
variables.

Built with Intel and Nag.




git-svn-id: truchas/trunk@202