Pass a pointer instead of an int for an array; closes #209
Previously, an array value was passed for what should have been a pointer to an
array value. Thanks to the previous commit, the compiler now warns about it:

/var/tmp/certik/truchas/build/src/packages/danu/src/python/GMVPYTHON_wrap.c(3812): warning #167: argument of type "int" is incompatible with parameter of type "void *"
                gmvwrite_cell_type("ptet4", 4, list[1]);
                                               ^

This then causes to segfault at runtime, because the value of the pointer was
the integer value (somerthing like "4"), which was an incorrect pointer.

We now pass a pointer to the array value, thus the pointer is valid and things
work at runtime.

git-svn-id: truchas/trunk@381