Change partitioning of the mesh served by legacy_mesh_api
This configures legacy_mesh_api to use the same partitioning of cells
and nodes used by the new mesh. Gap cells are tacked onto the end of
the list of (local) cells, so that the first part is 1-1 with the
on-process new mesh cells, and the second part is 1-1 with links that
came from gap cells.

Few tests must be modified to pass. This is done in subsequent commits.