Eliminate node data rearrangement from NN_GATHER_IMPL
The node ordering/partitioning presented by legacy_mesh_api coincides
now with that of the mesh, making the rearrangement of node-based data
unnecessary.  This eliminates that step from NN_GATHER_IMPL.  A side
effect is that the neighbor node lists in NGBR_VERTEX_ALL are ordered
differently (order here is insignificant).  This results in small
numerical roundoff differences (order of operation) that the solid
mechanics contact algorithm is extremely sensitive to, however.  This
commit includes a tweak to one test problem input so that it passes.