Eliminate node data rearrangement from VERTEX_IMPL
The node ordering/partitioning presented by legacy_mesh_api coincides
now with that of the mesh, making the rearrangement of node-based data
unnecessary.  This eliminates that step from NN_GATHER_IMPL.  A side
effect was a change in the way new gap nodes were handled (summing
node results to the parent (gnmap) -- which it shouldn't have been
doing in the first place since the legacy API now has exactly the
same nodes as the mesh).  This required updating the htvoid2 and
htvoid3 test problems; see issue #247 for background on the issue.