Eliminate the gap node map from the legacy mesh API.
This is no longer needed now that the legacy mesh API serves the same
nodes as the mesh.  Previously some internal interfaces in the mesh
were not present in the legacy mesh and the gap node map provided the
means to stitch the interface back together.  This (GNMAP) should not
have been used any longer, but still was by MESH_IMPL.  This has been
fixed, which also required some tweaks to the htvoid3 test to make it
pass again.