Refactor the cmake build system (resolves #71)
This is a major reworking of the cmake build system whose primary purpose
is to make the TPL builds and Truchas build entirely separate steps.  The
Truchas cmake configuration now expects to find suitable TPL installations;
it will not build any that are missing.  TPLs must be built and installed,
if necessary, prior to configuring the Truchas build. The cmake superbuild
system found in the new "tpl" subdirectory provides one means for doing
this.

In the process, the cmake build system was updated (3.5 is now required)
and greatly simplified.  A NetCDF that includes the netcdf-4 format (HDF5)
is now required, and in parallel, a parallel version of HDF5 is required.
These will be needed for upcoming capability enhancements.  The Cray Linux
Environment on Trinitite is also now supported.