Add support for large problems to view factor tools
This adds limited support for large problems to the RadE tool set
(genre, vizre, cmpre).  The size of the view factor matrix is now
practically unlimited (2^63 non-zeros).  But the mesh size and size
of the local portion of the view factor matrix are still limited to
a representation using 32-bit integers.