Improve incomplete BC error message (issue #115)
This adds the list of possible link set IDs to the error message written
when BC for heat transport/species diffusion are missing from part of the
boundary, together with the count of such boundary faces not belonging to
any face set or link set.  Additional diagnostic info on the latter faces
is written earlier when the mesh is instantiated.  Also fixes a typo bug
in global_ior (for bitfield type) that could result in an incomplete list
of set IDs.