Add legacy_matl_adapter class and extensions (not yet used)
This abstracts the procedures from legacy_matl_api as an abstract class
in order to facilitate exchanging one implementation for another. This
also includes two implementations: one using the original matl code (and
additions from legacy_matl_api), and another that uses a simple rank-2
array to hold the volume fractions for all materials and cells. This
new code compiles, but is not yet used.