Absorb flow_mesh components into unstr_mesh
This is a fairly invasive change. The fcell, cell_centroid, and
face_centroid arrays have been moved into the unstr_mesh type and the
flow_mesh objects simply replaced by unstr_mesh objects (and all the
%mesh%mesh stuff eliminated). An significant change to be aware of is
the order of the cells in the fcell array is swapped the face normal
points out of the first cell and into the second.