Add a static compressed "matl" implementation
The adds a compressed array implementation of the legacy_matl_adapter
class. In this implementation the volume fractions are stored in a
ragged array. Unlike the original matl, this structure is static.
The legacy_matl_api is configured to use this implementation in this
commit. NB: It is uncertain whether this one is working.