Selected genre code modernization
This is a preliminary step in applying an old MR for moving enclosures.
That MR included some code modernization updates that are pulled out
here: use of allocatable strings where appropriate, and primarily the
replacement of one-off structs to hold namelist data with parameter
lists, which pave the way for replacing namelist input with JSON input.