Changed and new genre inputs
Instead of listing the side set IDs of enclosure surfaces to displace
(`displace_side_set_ids`), one now specifies the element blocks to
displace (`displace_block_ids`). This is more natural from a user's
perspective. Associated with this is a new ENCLOSURE input variable
`exodus_block_modulus` which has the same purpose as the Truchas MESH
variable of the same name. This is important for mixed element meshes.