Code changes and instruction build on mac
- chaparral required a minor build system change:
Compilers running on apple already define __APPLE__ so there is no
reason to also define `macos`.  This extra definition was not
consistent with the compiler and lead to errors

- runinfo.c was doing unchecked string concatenation.  The strings
returned by the hostname calls are longer on mac.  This led to some
stack smashing errors.  Fixed with `snprintf`

- python testing libraries now look for shared libraries with
CMAKE_SHARED_LIBRARY_SUFFIX rather than just '.so'

- fixed make_directory_hier_c

- remove cpp dependence from pgslib since it didn't work with clang

- new cmake configuration for mac/nag