Add time-dependent view factor capability
This adds a capability for doing radiative heat transfer in a changing
enclosure that targets a directional and single crystal investment casting
process in which the filled mold is slowly withdrawn from the furnace to
control the solidification. The time-dependent view factors are linearly
interpolated from a time sequence of static view factors for different
enclosure configurations.