Update tabular solid fraction evaluation
This replaces an assertion that the computed solid fraction is in [0,1]
with code that simply clips the solid fraction to that interval. This
avoids unwanted assertions in occasional cases where the solid fraction
was outside the interval by just a few ulp. A more carefully crafted
assertion to check for true errors should probably be added back in.