Test the binary tarball at the CI
Currently the tarball is built, but not tested. As a result it is
actually broken currently, the Truchas binary is not included. We will
fix then in subsequent commits. This commit adds a test to unack the
tarball and run the Truchas binary, in a stock Ubuntu image, which
ensures that the binary works across linux distributions (it is built in
Centos 7), without Intel compilers present.

Later on we should also add a simple test example to run. Until then we
just run Truchas with "-h" to print help, which still tests that all the
dynamic libraries are present and work.