New formulation for solidification velocity (#338)
This ditches the complex approach to solidification velocity that uses
the time rate of change of solid fraction and mimetic reconstruction
of the solid fraction gradient, in favor of a simpler and far less
costly indirect approach that uses the time rate of change of temperature
and thermal gradient which is already available. The two formulations are
identical in the continuous context, but not necessarily so in the
discrete.