WIP: Add gap contact BC type
This adds a gap-contact-like condition that enforces zero-displacement across
gaps. There is still an outstanding bug that prevents this from converging on
a rotated version of the stretch-gap-x input file, but the solution on the
aligned version looks like the expected result for the condition imposed.

The bigger problem right now is this zero-displacement condition isn't right at
gap contacts. The condition should be zero-penetration, so that pushing force is
transmitted across gaps, but pulling force isn't (to an extent). There also must
be special operation changes at corners, so corner gaps disallow penetration in
multiple directions.