WIP: Prototype HT solver using vector-class-based DAE integrator (#450)
This adds a new HT solver that uses a redesigned DAE integrator that uses
abstract vectors rather than intrinsic rank-1 real arrays. This pure HT
solver is used in place of the HTSD solver whenever species diffusion is
not enabled.

This is working for all HT and HT+flow test problems that had used the
HTSD solver.