WIP: Write new contact type
The new contact type is now "working". A new test is added stretch-contact to
test it, which is still underway. But the 1-contact BC is now producing correct
results, and running correctly in parallel (which is more than I can say for the
last contact type). However, performance seems to have degraded. More iterations
are required to solve systems in this commit than the last commit, even when
contact is not present. For instance, look at the stretch-nx-2 problem, which
now requires ~600 iterations. I think the previous commit required ~300
iterations, and the legacy code (stretch-nx-2) takes <200 iterations. The new
code produces the same result as the old code, so for some reason the
preconditioner is of poor quality, and even poorer quality as of this commit,
though I'm not sure what I actually changed.