Truchas will not overwrite its output data
To restore the previous behavior run with the new `-f` flag.

The check for existing output data files is done on `.h5` output files
as well as `.log` files.

The new global boolean, `overwrite_output`, in truchas_env governs
whether overwriting of data is allowed or if it will cause a fatal
error.  The non-python based tests were modified to explicitly set
this to true.

Added a new environment variable `TRUCHAS_OVERWRITE_OUTPUT` that is
used by the python truchas wrappers to call truchas with the `-f`
flag. This allows for simply adding the environment variable in our
AddPytruchasTest rather than modifying all the tests