Drop delayed initialization of index_map components
The data used by the type-bound distribute and collate subroutines
was delayed until the first time one of those procedures was called.
This had the unpleasant side effect of requiring index_map arguments
to be intent(inout) in cases when they would otherwise be intent(in),
and for no readily apparent reason. In practice there was no advantage
to delaying the initialization.