Drop gap elements from the legacy mesh
The legacy mesh had transformed interface links to gap elements
and appended them to the list of on-process cells. This commit
omits that step. Note that some of the no-longer-used components
of the legacy API may no longer be completely valid due to this
change, but all tests are passing.

With this change the legacy mesh ncells value is the same as the
ncell_onP mesh value, facilitating migration to the mesh structure.