Workaround gfortran 11.2 ICE
The preceding change to EM_permittivity, etc., triggered an ICE in
the gfortran 11.2 compiler; not in those functions themselves, but
in code that called them. This commit works around the ICE by
converting these array-valued functions to subroutines where the
array is passed as an argument (a better design anyway).