Add microwave modeling capability
This is a major update to the electromagnetic modeling capability, and
is the squash of 176 commits made between Apr 7 2023 and Sep 15 2025.
At the core is a new frequency domain EM solver with options for iterative
and direct linear solvers. This is the basis for a new microwave heating
simulation physics option and a time-harmonic electrostatics simulation
tool. In addition it provides a FD solver alternative to the existing
time domain EM solver used for induction heating simulations.

Co-authored-by: Neil N. Carlson <neil.n.carlson@gmail.com>