Refactored emfd_nlsol_solver_type.F90
This maintains the overall structure between parts of the FD solver,
but significantly rewrites the internals. The system matrix was
a single PCSR matrix for the double sized system (real, imaginary)
parts. This is replaced by a 2x2 block system according to the
partition of the unknowns between real and imaginary parts, where
each block is a separate PCSR matrix all having the identical
structure. This *greatly* simplifies the initialization of the
system, allowing much of the work to be off-loaded to existing
capabilities provided by the pcsr_matrix/pcsr_graph types and
upper_packed_matrix_proc procedures.

This also eliminates the em_bc_type module which was a thin
container for the ebc and hbc BC objects.