Refactor Hiptmair preconditioner
This makes use of a new block CSR matrix type to store the matrices
as a CSR matrix with 2x2 block elements corresponding to the real
and imaginary unknown parts. This reduces the graph storage by 4x
and greatly simplifies the initialization of the matrices. The now
modified GS relaxation procedure is moved into the bcsr_matrix_type
module. An alternative storage scheme was also added (along with its
associated apply procedure) which stores the matrices as a 2x2 block
system partitioned according to the real/imaginary components. Each
block is an edge (or node) based CSR matrix. This also reduces the
graph storage by 4x, but also would enable the element storage to
be reduced by 2x because there are really only 2 distinct blocks.
However GS on this ordering of unknowns is much less effective.

These changes break the AMS solver; repairs are needed.

Still not workin in parallel.