Vector-based FD solvers
This modifies the generic gmres and nlk solvers to operate on
abstract vector objects in preparation for moving to parallel.
This approach allows the use of unknown arrays that include
off-process elements that are hidden to the solvers, and is
an alternative to copying unknown data between arrays that only
include on-process elements (operated on by solvers) and extended
arrays that include off-process elements that are needed to form
the equations.