Improved quadrature of EM boundary data
The boundary face integrals of the magnetic field data was computed
by first interpolating the source field into the edge space (circulation
on edges using trapezoid rule) and then analytically computing the
integrals with that interplated function. This commit replaces that
with a much more accurate direct approximation of the integrals using
7-point Gaussian quadrature with the primitive source field. This
quadrature is 5th order accurate (4th order in the source field).