Add specification of "free-space" nodes for AMS
This re-enables and corrects the specification of nodes in the
interior of the free-space region (i.e., the region without
dissipation). Test problems (not tied into ctest) were added
in test/em5 that include free-space in the EM domain.

This also adds output of the complex E and B phasors to the
VTK HDF5 output file.