Additions to the AMS preconditioner
This adds the computation of the Aalpha and Abeta Poisson matrices
that are optional inputs to the AMS solver. The calls to pass them
to Hypre are currently commented out because the GMRES diverges if
they are used.

This also adds a call to HYPRE_AMSSetProjectionFrequency (hardwired
frequency value). This too is commented out, but some values give
more rapid convergence and more accurate free-space solutions (visual
inspection), while others lead to non-convergence, without any
apparent pattern. The AMS documentation also describes the alternative
HYPRE_AMSProjectOutGradients which is added after the solve call,
but this segfaults and is commented out.

Finally, scaling was removed from the E field.