Option to use AMS for TD solver
This adds hypre_ams_type.F90 which provides an interface to
Hypre's AMS solver, and modifies the TD solver to allow its
use as an alternative to the existing CG solver.

This also massages the ELECTROMAGNETICS namelist input variables
to reduce the proliferation of specifically-named variables for
all the various solver options that now exist by settling on
common names where appropriate. (RM needs to be updated.)