Add Boomer option for TD CG Hiptmair preconditioner
The original Hiptmair preconditioner for the TD CG solver used GS
sweeps as the smoother for both the original system and subspace
correction system. This adds an option to use BoomerAMG in place
of the GS smoothers. This works in both serial and parallel. In
limited testing it requires fewer CG iterations, but is slower
overall due to the greater cost of each iteration. At present it
uses Hypre defaults for Boomer.