Add CS-MINRES linear solver
This adds the option of using a minres linear solver for the frequency
domain EM model. It currently uses no preconditioner. It replicates
the results for the linear system of the em5/emfd.inp test problem
(em-qcyl-tet-large mesh) obtained in a standalone driver for the solver,
and works in parallel.

Convergence is very slow without a preconditioner, and is not yet really
viable. It will require a different precondition than is currently
available.