Complexified fdme_model_type
This adds a complex parallel CSR matrix type and refactors
fdme_model_type to generate this type of matrix. This eliminates
two vector copies for each matvec that were needed to use the
original real vector based matvec. The original functionality
is still there for other solvers. This also adds diagonal
preconditioning for the CS-MINRES solver (hardwired for now).