Add GS preconditioner for FD minres solver
This adds a symmetric GS preconditioner (hardwired) for the fd minres
solver. It applies GS to the real part of the matrix (symmetric, though
not necessarily positive definite). In serial it greatly reduces the
number of iterations over diagonal preconditioning, and the overal cost
is significantly reduced as well.

This also "works" in parallel, though I'm unconvinced that the way it
is being used in parallel is truly correct and requires further work.