Add abstract vector based cs_minres
This adds a version of cs_minres that operates on an abstract complex
vector type. A version of fdme_cs_minres_solver that uses it was added
(temporarily?) in order to test and verify it. Additionally there was
a lot of refactoring to push the (real) vector objects down inside
the various fdme_*_solver types where it belongs, so that all the
higher level code make direct use of complex arrays.