Fix bug in refine_hex_to_tet
Fixes an error which resulted in neighboring hex cells splitting the
shared face in different ways. The fixed algorithm still assumes a
certain compatiblity in the orientation of the neighboring hex cells
that is satisfied by create_rect_exodus_mesh satisfies, but which will
not be true of general hex meshes. For that reason it has been moved
from the exodus_mesh_tools module to the exodus_mesh_factory module
with the appropriate warnings.