Add Boomer as a preconditioner for minres
This adds BoomerAMG as an alternative option to block Gauss-Seidel
preconditioning for the minres solver for FD EM.

This is just a basic implementation. No boomer parameters can be set
and only a single cycle is used; improvements will follow.

It works, but not necessarily better than GS, though it looks like
it will scale better.