Add minres preconditioning matrix option
This adds the input variable minres_pc_damping which modifies the matrix
the minres preconditioner is based on. By default it is the real part of
the complex symmetric system matrix Re{A} = A1 - A2 which is indefinite.
Here A1 is the curl-curl matrix. Instead the the preconditioner is based
on the damped matrix Re{A} + minres_pc_damping * A2. When the variable
is 1, the matrix is simply A1 which is positive semi-definite, and when
> 1 the matrix is positive definite. The default is no damping, Re{A}.

This also extracts the status flag from the CS-MINRES solver and adds
a check for non-positiveness of the preconditioner.