Rewrite of cs_minres_solver
This is a complete rewrite of the abstract vector based CS-MINRES
solver which strips out the extra "QLP" process, and other aspects,
not necessary for our application. It also replaces the fdme_zvector
type with a generic imap_zvector type that is more generally useful.
All solvers using CS-MINRES were migrated to this vector-based version
and now-unused related code deleted.