Add multicomponent alloy solidification
This adds a prototype capability to simulate the solidification of
multicomponent alloys. Two models are implemented: the lever rule
and the Wang-Beckermann back-diffusion model with parameter that
spans between the Scheil model and asymptotically toward the lever
rule. This is based on a eutectic-type phase diagram with linear
approximations to the liquidus and solidus surfaces.

This type of simulation is enabled using the `alloy_solidification`
keyword of the PHYSICS namelist, which implies the `heat_transport`
keyword. There is a new namelist ALLOY namelist that specifies the
parameters associated with the solidification model and phase
diagram. The model is compatible with fluid flow, and when enabled
will include advection of enthalpy and liquid solute species.

This prototype implementation is subject to a number of limitations:
* It is not compatible with any physics model other than flow.
* The thermal BC are limited to Dirichlet, flux, and external HTC
  conditions, and does not support internal interfaces and interface
  conditions.
* Only a single 2-phase material -- the alloy -- is supported (and
  no VOID). The usual phase change description of material is
  required but only for the purposes of defining the volumetric
  enthalpy of the two phases; the actual phase change (e.g. liquid
  fraction) is governed by the new solidification models.
* Simulation restarts are not implemented.

There are several new integrated tests in the `test/alloy` directory.
In addition there are several low-level drivers for the solidication
models in `heat_species_transport/Test`; these are compiled but not
integrated into CTest and must be run manually.