Fork ht_* code to alloy_*
This makes a copy of the ht_* solver code to alloy_*.  It strips
handling of void and viewfactor radiation from the forked code so
that it is pure heat transfer. The new solver code path is used
when the PHYSICS namelist variable alloy_solidification is enabled
(and the problem characteristics are compatible). There is a new
test problem alloy1 that solves a 1D phase change problem.

This serves as the starting point for development of a multicomponent
phase change (and micro/macrosegregation) model.