New material methods
This adds a couple of capabilites to the materials component:
* A single phase material can be created from a phase of a
  multicomponent material.
* A single phase material property can be created from one
  phase of a multicomponent material.

These were intended to be used by the multicomponent alloy
solidification model, but weren't in the end (so far). Neither
have been tested properly, and I suspect there is something
not right with the second item when it comes to the finite
difference computation of the derivative.