Add liquid fraction constraint
This adds the liquid fraction relation as an algebraic constraint
in the DAE system. This hacked-in version uses the smooth phase
change funtion for the liquid fraction. The alloy2 test (same as
alloy1?) recovers the reference alloy1_golden solution to high
accuracy. (Though the tests fail due to very tight tolerences.)
There is no significant difference in the number of steps (and
execution time).