Add lever rule phase change
This modifies the algebraic liquid fraction equation to implement
the lever rule corresponding to a multicomponent phase diagram.
The concentration is a uniform constant (no flow, no diffusion).
This is working for concentrations that do not end in a eutectic
transformation. For a eutectic transformation, the original plan
will not work, and that transformation will need to be spread out
over a small temperature range.