Mollify the eutectic transformation
This spreads out the isothermal transformation that occurs at
the eutectic temperature over a temperature interval. For an
isothermal transformation dH/dT would be infinite. The approach
taken here chooses a temperature interval so that dH/dT is a
prescribed (large) value (currently hardwired in multicomp_lever_type).
The larger the value, the smaller the interval (and more costly the
computation). This is working (see alloy4.inp problem)