Add back-diffusion eutectic
This adds the eutectic transformation to the back-diffusion model,
but it is not working yet. It runs into singular jacobian issues
when exiting the eutectic transformation. See the uniform.inp
test which mirrors the single-cell standalone test driver (which
didn't fail).

This also enables solute advection for the lever model, which already
had a working eutectic transformation. This seems to be working okay.