Density varies with concentration; lever rule updates
This adds the solute concentration in liquid to the external state
that flow sees along with temperature, which allows the density_delta
function to depend on solute concentrations as well as temperature.
Currently that function must be specified directly as a function of
(T,C1,...,Cm) and not in terms of solutal expansion coefficients.

This also adds the final missing piece needed for solute advection
when using the lever rule alloy model. This is working.

It also changes the algorithm for the eutectic transformation,
and specifying the temperature interval over which it occurs,
rather an obscure max of dH/dT.